+++ crates/module-system/module-implementations/module-template/src/event.rs
@@ -7,6 +7,7 @@
     Debug,
     PartialEq,
     Clone,
+    schemars::JsonSchema,
 )]
 #[serde(rename_all = "snake_case")]
 pub enum Event {