+++ crates/module-system/module-implementations/sov-attester-incentives/src/event.rs
@@ -12,6 +12,7 @@ use crate::SlashingReason;
     Debug,
     PartialEq,
     Clone,
+    schemars::JsonSchema,
 )]
 #[serde(rename_all = "snake_case")]
 pub enum Event<S: Spec> {